+++ Changelog.md
@@ -1,3 +1,8 @@
+### v1.9.105 (24-Jun-2023)
+bây giờ sau khi lấy danh sách video của 1 kênh Youtube thì:
+- có thể mở file kết quả trực tiếp bằng Excel (không cần cài LibreOffice nữa)
+- có thể sắp xếp theo lượt xem hoặc ngày đăng 
+
 ### v1.9.104 (21-Jun-2023)
 - thêm chức năng lấy danh sách video của 1 kênh Youtube
 - thêm đường dẫn đến trang lấy hình avatar và banner của 1 kênh Youtube (https://imageyoutube.com/?)